Add print Tx/Rx serial stream arg to app without setting Trace
More than a few times I have needed a way to see what was being sent
to and from RCP.  Before this patch the only way I could do this was
to enable Trace level logging (which is not easy to do unfortunately).
But the problem with this is that if I enable Trace level debugging, I
get TONS of logs.  Metric TONS.

That is not helpful to me.  So instead I wrote this patch that will
allow me to extract out the Tx/Rx data by passing the following arg

./main.py -- --print_tx_rx